Fix inter- and intra-module pagination order
- Between modules, `/state/` articles were coming before `react` ones
for some odd reason related to how `gatsby-source-filesystem` loads
files into graphql.
- Within modules, order was like `1.md`, `10.md`, `2.md` because gatsby
was loading files in lexicographic order. It wasn't *exactly* in that
order either, because no `sort` criterium was being passed, so it was
almost random.

Renaming the files to `01.md`, etc. (left zero padding, previous commit)
and using that as the order (this commit) solved both problems.