Fix and bulletproof page path inconsistencies
Gatsby's createPage expects a `path` field that must start with `/`.

The `slug`-related variables were spread all over and being used only
for paths.

Since paths require a `/` at the beginning, the "prepend a `/` at the
beginning" implementation detail was being repeated everywhere, and
forgetting it caused a bug.

When prepending is forgotten, as in `CurriculumLink`, urls become
relative, and a slug like `state/title` leads to a 404 url like
`/state/state/title`.

This wasn't caught before because, by luck, on `/` because the relative
and absolute paths coincided.

This fixes that and hides the implementation detail.